Some work towards composable/customizable styles.
Replace most direct method calls on message objects with calls to
methods on the style that (by default) just delegate to the object.

This allows for subclasses to override these methods to customize the
display. With the use of the C3 mro (use mro 'c3'; in perl 5.10) in
subclasses, this even allows for subclasses to be composed with each
other by making a subclass with multiple parent classes.